requirements.txt
# ============================================================
# APEX INDIA — Autonomous Quantitative Trading System
# Python Dependencies (pinned for reproducibility)
# ============================================================

# ── Core ─────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz==2024.2

# ── Technical Indicators ─────────────────────────────────────
pandas-ta==0.3.14b1

# ── Machine Learning ─────────────────────────────────────────
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.3
statsmodels==0.14.4
optuna==4.1.0

# ── Deep Learning (PyTorch — CPU by default) ─────────────────
torch==2.5.1
transformers==4.46.3
stable-baselines3==2.4.0

# ── NLP & Sentiment ──────────────────────────────────────────
beautifulsoup4==4.12.3
feedparser==6.0.11
requests==2.32.3
lxml==5.3.0

# ── Broker APIs ──────────────────────────────────────────────
kiteconnect==5.0.1

# ── Databases ────────────────────────────────────────────────
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
influxdb-client==1.47.0
alembic==1.14.0

# ── Async & WebSockets ───────────────────────────────────────
asyncio-mqtt==0.16.2
aiohttp==3.11.9
websockets==14.1

# ── Dashboard & Visualization ────────────────────────────────
streamlit==1.40.2
plotly==5.24.1

# ── Alerts & Notifications ───────────────────────────────────
python-telegram-bot==21.7

# ── Scheduling ───────────────────────────────────────────────
APScheduler==3.10.4

# ── Logging ──────────────────────────────────────────────────
loguru==0.7.3

# ── Backtesting ──────────────────────────────────────────────
vectorbt==0.26.2

# ── Utilities ────────────────────────────────────────────────
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
tqdm==4.67.1
python-dotenv==1.0.1
